requirements.txt
langgraph>=0.6.4
langchain>=0.3.27
langchain-openai>=0.3.29
langchain-community
langchain-chroma
python-dotenv>=1.1.1
beautifulsoup4
requests
pypdf
chromadb
lxml
reportlab

